refactor: align map and utility file naming
